sidebar selection cleaned up
