+++ tfbpshiny/modules/select_datasets/server/__init__.py
@@ -1,0 +1,11 @@
+from tfbpshiny.modules.select_datasets.server.sidebar import selection_sidebar_server
+from tfbpshiny.modules.select_datasets.server.workspace import (
+    select_datasets_server,
+    selection_matrix_server,
+)
+
+__all__ = [
+    "select_datasets_server",
+    "selection_sidebar_server",
+    "selection_matrix_server",
+]
